Prevent save and load from being performed on uninitialized clothes (since the vertex count remains zero)
The cloth data is sometimes loaded or saved on an uninitialized cloth component, which has a "vertices" count of zero. This is problematic during export because this generates an asset of 0 vertices, which then causes an array error when loaded on the target cloth.

Therefore, prevent both saving and loading if we detect that the vertex count is 0 on the cloth component.

Closes #6